requirements.txt
deepface
flask
flask-cors
pillow
numpy